requirements.txt
Django==1.10.6
djangorestframework==3.5.4
psycopg2>=2.7,<3.0
requests==2.9.1